library function integration for chat
