+++ packages/messenger-widget/src/components/ConfigureProfileBox/ConfigureProfileBox.css
@@ -1,7 +1,7 @@
 .box-outer-container {
     padding: 10px 1px 10px 0px;
     bottom: 1%;
-    margin: 1rem 1rem 1rem 1rem;
+    margin: 2rem 1rem 1rem 1rem;
 }
 
 .configure-msg-box {
